requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
huggingface-hub>=0.20.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.9.0
streamlit>=1.28.0

